changed tuple size to 16 bytes and array allocations to be 50% more than expected

+++ MiniProject1/Independent_Partitioning_Algorithm.h
@@ -8,8 +8,8 @@ using namespace std;
 
 class Independent_Partitioning_Algorithm {
 public:
-    void IndependentPartition(vector<tuple<int,int> > tuples, int threadCount, int hashBits);
-    static void IndependentProcess(vector<tuple<int,int> > tuples, int from, int to, int hashBits);
+    void IndependentPartition(vector<tuple<long long, long long> > tuples, int threadCount, int hashBits);
+    static void IndependentProcess(vector<tuple<long long, long long> > tuples, int from, int to, int hashBits);
 };
